requirements.txt
Flask
Flask-Bcrypt
cryptography
pyjwt
